Separate out frontend token tests
